Prefer named exports & components
= Use named exports instead of default
= Use components instead of directives and controllers
