+++ components/add-layers/capabilities-error.component.ts
@@ -1,0 +1,6 @@
+export const HsGetCapabilitiesErrorComponent = {
+  template: require('./partials/dialog_getcapabilities_error.html'),
+  controller: function (scope, element, attrs) {
+    scope.capabilitiesErrorModalVisible = true;
+  },
+};
