+++ components/add-layers/nested-layers-table.component.ts
@@ -1,0 +1,7 @@
+export const HsNestedLayersTableComponent = {
+  template: require('./partials/nested-layers-table.directive.html'),
+  scope: {
+    layers: '=layers',
+  },
+  controller: ['$scope', function ($scope) {}],
+};
