--- components/add-layers/nested-layers-table.directive.ts
@@ -1,13 +1,0 @@
-/**
- * @param HsConfig
- */
-export default function (HsConfig) {
-  'ngInject';
-  return {
-    template: require('./partials/nested-layers-table.directive.html'),
-    scope: {
-      layers: '=layers',
-    },
-    controller: ['$scope', function ($scope) {}],
-  };
-}
