+++ components/add-layers/resample-dialog.component.ts
@@ -1,0 +1,6 @@
+export const HsResampleDialogComponent = {
+  template: require('./partials/dialog_proxyconfirm.html'),
+  controller: function (scope, element, attrs) {
+    scope.resampleModalVisible = true;
+  },
+};
