+++ components/add-layers/wms/add-layers-wms.service.ts
@@ -20,7 +20,7 @@ import {getPreferedFormat} from '../../../common/format-utils';
  * @param HsUtilsService
  * @param HsConfig
  */
-export default function (
+export const HsAddLayersWmsService = function (
   $rootScope,
   HsMapService,
   HsWmsGetCapabilitiesService,
@@ -412,4 +412,4 @@ export default function (
   };
 
   return me;
-}
+};
